+++ src/main/java/com/moment/the/exception/legacyHandler/AccessNotFoundHandler.java
@@ -1,13 +1,12 @@
-package com.moment.the.exception.handler;
-
+package com.moment.the.exception.legacyHandler;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-public class AnswerAlreadyExistsHandler {
+public class AccessNotFoundHandler {
     public void handle(HttpServletRequest req, HttpServletResponse res) throws IOException, ServletException {
-        res.sendRedirect("/exception/answer-already-exists");
+        res.sendRedirect("/exception/access-not-found");
     }
 }
