+++ src/main/java/com/moment/the/exception/legacyHandler/NoCommentHandler.java
@@ -1,4 +1,4 @@
-package com.moment.the.exception.handler;
+package com.moment.the.exception.legacyHandler;
 
 import lombok.extern.slf4j.Slf4j;
