+++ src/main/java/com/moment/the/exception/legacyHandler/UserAlreadyExistsHandler.java
@@ -1,4 +1,4 @@
-package com.moment.the.exception.handler;
+package com.moment.the.exception.legacyHandler;
 
 import lombok.extern.slf4j.Slf4j;
 
@@ -8,8 +8,8 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 @Slf4j
-public class UserNotFoundHandler {
+public class UserAlreadyExistsHandler {
     public void handle(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
-        response.sendRedirect("/exception/userAlreadyExists");
+        response.sendRedirect("/exception/userNotFound");
     }
 }
